backend/pyproject.toml
[project]
name = "chatkit-backend"
version = "0.1.0"
description = "FastAPI backend service for ChatKit boilerplate"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.1,<0.116",
    "uvicorn[standard]>=0.36,<0.37",
    "openai>=1.40",
    "openai-chatkit>=1.4.0,<2",
    "jinja2",
    "httpx",
    "python-dotenv",
    "x402",
    "eth-account",
    "coinbase-agentkit>=0.7.4",
    "cdp-sdk>=1.34.0",
    "nest-asyncio>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.4,<0.7",
    "mypy>=1.8,<2",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]
